add support for APS style arXiv identifiers
While working on Inspire we have seen cases where
arXiv identifiers are in the mixed old- and new-style
form, where the identifier is of the new style, but
is prefixed with the class, like the old style IDs,
e.g. arXiv:hep-th/1601.07616. This brings support
for such identifiers to idutils.

Signed-off-by: Szymon Łopaciuk <szymon.lopaciuk@cern.ch>